% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day10_input}
\alias{day10_input}
\title{Input for Day10 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day10_input}}{

A single string containing one CPU instruction per line
}
}
\source{
\url{https://adventofcode.com/2022/day/10}
}
\usage{
day10_input
}
\description{
Input for Day10 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
