% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day11_input}
\alias{day11_input}
\title{Input for Day11 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day11_input}}{

A single string containing monkey descriptions separated by empty lines
\describe{
\item{Starting items}{A list of integers indicating your initial worry for each items the monkey holds}
\item{Operation}{A formula describing how your worry about an item changes when the monkey inspects it}
\item{Test}{An if-else expression describing to which monkey the item will be passed based on your worry about the item}
}
}
}
\source{
\url{https://adventofcode.com/2022/day/11}
}
\usage{
day11_input
}
\description{
Input for Day11 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
