% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day12_input}
\alias{day12_input}
\title{Input for Day12 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day12_input}}{

A single string containing a topographic map. Each character represents a height level from 'a' = 0 to 'z' = 25. Moreover, the start point is 'S' at level 0 and the end point is 'E' at level 25. The distance between adjacent characters is 1.
}
}
\source{
\url{https://adventofcode.com/2022/day/12}
}
\usage{
day12_input
}
\description{
Input for Day12 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
