% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day13_input}
\alias{day13_input}
\title{Input for Day13 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day13_input}}{

A single string containing pairs of packets separated by an empty line. Each non-empty line is a packet, which is a comma-separated list of other packets and integers enclosed in square brackets
}
}
\source{
\url{https://adventofcode.com/2022/day/13}
}
\usage{
day13_input
}
\description{
Input for Day13 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
