% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day14_input}
\alias{day14_input}
\title{Input for Day14 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day14_input}}{

A single string containing one description for a rock formation per line. Each description is a sequence of corners of the rock formation.
}
}
\source{
\url{https://adventofcode.com/2022/day/14}
}
\usage{
day14_input
}
\description{
Input for Day14 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
