% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day15_input}
\alias{day15_input}
\title{Input for Day15 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day15_input}}{

A single string containing one sensor position together with the closest beacon position per line.
}
}
\source{
\url{https://adventofcode.com/2022/day/15}
}
\usage{
day15_input
}
\description{
Input for Day15 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
