% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day16_input}
\alias{day16_input}
\title{Input for Day16 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day16_input}}{

A single string containing one valve specification per line. Each shows the name, the possible release rate and the other valves the valve is connected to.
}
}
\source{
\url{https://adventofcode.com/2022/day/16}
}
\usage{
day16_input
}
\description{
Input for Day16 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
