% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day17_input}
\alias{day17_input}
\title{Input for Day17 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day17_input}}{

A single string containing in which each character represents the direction of a jet.
}
}
\source{
\url{https://adventofcode.com/2022/day/17}
}
\usage{
day17_input
}
\description{
Input for Day17 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
