% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day18_input}
\alias{day18_input}
\title{Input for Day18 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day18_input}}{

A single string containing one point in three dimensions on each line
}
}
\source{
\url{https://adventofcode.com/2022/day/18}
}
\usage{
day18_input
}
\description{
Input for Day18 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
