% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day19_input}
\alias{day19_input}
\title{Input for Day19 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day19_input}}{

A single string containing blueprint per line
}
}
\source{
\url{https://adventofcode.com/2022/day/19}
}
\usage{
day19_input
}
\description{
Input for Day19 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
