% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day1_input}
\alias{day1_input}
\title{Input for Day1 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day1_input}}{

A single string containing several blocks of lines separated by blank lines.
Each non-empty line contains an integer.
}
}
\source{
\url{https://adventofcode.com/2022/day/1}
}
\usage{
day1_input
}
\description{
Input for Day1 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
