% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day20_input}
\alias{day20_input}
\title{Input for Day20 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day20_input}}{

A single string containing one integer per line
}
}
\source{
\url{https://adventofcode.com/2022/day/19}
}
\usage{
day20_input
}
\description{
Input for Day20 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
