% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day21_input}
\alias{day21_input}
\title{Input for Day21 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day21_input}}{

A single string containing one monkey description per line
}
}
\source{
\url{https://adventofcode.com/2022/day/21}
}
\usage{
day21_input
}
\description{
Input for Day21 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
