% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day22_input}
\alias{day22_input}
\title{Input for Day22 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day22_input}}{

A single string containing a map with empty spaces as '.' and walls as '#' separated from a list of move instructions by an empty line
}
}
\source{
\url{https://adventofcode.com/2022/day/22}
}
\usage{
day22_input
}
\description{
Input for Day22 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
