% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day23_input}
\alias{day23_input}
\title{Input for Day23 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day23_input}}{

A single string containing a map with empty spaces as '.' and elves as '#'
}
}
\source{
\url{https://adventofcode.com/2022/day/23}
}
\usage{
day23_input
}
\description{
Input for Day23 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
