% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day24_input}
\alias{day24_input}
\title{Input for Day24 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day24_input}}{

A single string containing a map of a valley with blizzards. Empty spaces as are shown as '.', the boundaries as '#' and the blizzards as '<', '>', '^' or 'v' depending on the direction they move in
}
}
\source{
\url{https://adventofcode.com/2022/day/24}
}
\usage{
day24_input
}
\description{
Input for Day24 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
