% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day25_input}
\alias{day25_input}
\title{Input for Day25 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day25_input}}{

A single string containing a list of fuel requirements encoded in SNAFU, each on its own line
}
}
\source{
\url{https://adventofcode.com/2022/day/25}
}
\usage{
day25_input
}
\description{
Input for Day25 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
