% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day2_input}
\alias{day2_input}
\title{Input for Day2 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day2_input}}{

A single string containing one Rock Paper Scissors Strategy Guide item per line
\describe{
\item{OpponentPlay}{Opponent play encoded as 'A' = 'Rock', 'B' = 'Paper' and 'C' = 'Scissors}
\item{ResposnseCode}{Reponse code with values 'X', 'Y' and 'Z'; function not specified}
}
}
}
\source{
\url{https://adventofcode.com/2022/day/2}
}
\usage{
day2_input
}
\description{
Input for Day2 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
