% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day3_input}
\alias{day3_input}
\title{Input for Day3 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day3_input}}{

A single string containing the contents of one rucksack at a time
Each letter represents an item type. The first half of the items is in the left and the other half in the right compartment.
}
}
\source{
\url{https://adventofcode.com/2022/day/3}
}
\usage{
day3_input
}
\description{
Input for Day3 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
