% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day4_input}
\alias{day4_input}
\title{Input for Day4 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day4_input}}{

A single string containing one pair of section ranges per line.
Section ranges are presented as a pair of integers separated by a hyphen.
}
}
\source{
\url{https://adventofcode.com/2022/day/4}
}
\usage{
day4_input
}
\description{
Input for Day4 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
