% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day5_input}
\alias{day5_input}
\title{Input for Day5 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day4_input}}{

A single string containing two parts separated by an empty line.
The first section is a depiction of the initial stack configuration.
The second section is a list of moves to perform, one per line.
}
}
\source{
\url{https://adventofcode.com/2022/day/5}
}
\usage{
day5_input
}
\description{
Input for Day5 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
