% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day6_input}
\alias{day6_input}
\title{Input for Day6 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day6_input}}{

A single string of seemingly random characters representing a communication signal.
}
}
\source{
\url{https://adventofcode.com/2022/day/6}
}
\usage{
day6_input
}
\description{
Input for Day6 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
