% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day7_input}
\alias{day7_input}
\title{Input for Day7 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day7_input}}{

A single string of console output generated by browsing a file system.
Each command starts with a $ on a new line.
Lines not starting with a $ are command outputs.
}
}
\source{
\url{https://adventofcode.com/2022/day/7}
}
\usage{
day7_input
}
\description{
Input for Day7 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
