% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day8_input}
\alias{day8_input}
\title{Input for Day8 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day8_input}}{

A single string containing a square grid of tree heights.
}
}
\source{
\url{https://adventofcode.com/2022/day/8}
}
\usage{
day8_input
}
\description{
Input for Day8 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
