% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day9_input}
\alias{day9_input}
\title{Input for Day9 of Advent of Code 2022 for Max Doerner}
\format{
\subsection{\code{day9_input}}{

A single string containing one move of the form 'direction distance' per line, where U, D, L, R encode the directions
}
}
\source{
\url{https://adventofcode.com/2022/day/9}
}
\usage{
day9_input
}
\description{
Input for Day9 of Advent of Code 2022 for Max Doerner
}
\keyword{datasets}
