% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode_point}
\alias{decode_point}
\title{Decode Point}
\usage{
decode_point(point_text)
}
\arguments{
\item{point_text}{A character vector representing a point as generated by \code{encode_point}}
}
\value{
A double vector representing the point
}
\description{
Decode Point
}
