% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_point}
\alias{encode_point}
\title{Encode Point}
\usage{
encode_point(point)
}
\arguments{
\item{point}{A numeric vector}
}
\value{
A character vector representing the point
}
\description{
Encode Point
}
