% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{game_outcome}
\alias{game_outcome}
\title{Results of Playing Rock Paper Scissors}
\usage{
game_outcome(player_play, opponent_play)
}
\arguments{
\item{player_play}{A factor vector of Rock Paper Scissor plays}

\item{opponent_play}{A factor vector of Rock Paper Scissor plays the same length as \code{player_play}}
}
\value{
A factor with the result of the games with levels Loss, Draw and Win
}
\description{
Results of Playing Rock Paper Scissors
}
