% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\docType{data}
\name{game_outcomes}
\alias{game_outcomes}
\title{Result Table for Rock Paper Scissors Games}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
game_outcomes
}
\description{
Result Table for Rock Paper Scissors Games
}
\keyword{datasets}
