% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{input_parse_function_name}
\alias{input_parse_function_name}
\title{Name of the parse function for a specific day of AdventOfCode}
\usage{
input_parse_function_name(day)
}
\arguments{
\item{day}{An integer between 1 and 25 specifying the day of AdventOfCode}
}
\value{
Name of the input parse function for the specified \code{day}
}
\description{
Name of the parse function for a specific day of AdventOfCode
}
