% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{opponent_play_from_code}
\alias{opponent_play_from_code}
\title{Opponent Play from Code}
\usage{
opponent_play_from_code(code)
}
\arguments{
\item{code}{The character vector with items A, B or C}
}
\value{
A factor indicating the opponent play
}
\description{
Uses the Strategy Guide explanation to determine the enemy play
}
