% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{parse_day11_input}
\alias{parse_day11_input}
\title{Parsed Input to Day11 of Advent Of Code 2022}
\usage{
parse_day11_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 11 of AdventOfCode}
}
\value{
A list of monkeys
}
\description{
Parsed Input to Day11 of Advent Of Code 2022
}
