% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{parse_day12_input}
\alias{parse_day12_input}
\title{Parsed Input to Day12 of Advent Of Code 2022}
\usage{
parse_day12_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 12 of AdventOfCode}
}
\value{
A list containing the start point, the end point and a height map
}
\description{
Parsed Input to Day12 of Advent Of Code 2022
}
