% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{parse_day13_input}
\alias{parse_day13_input}
\title{Parsed Input to Day13 of Advent Of Code 2022}
\usage{
parse_day13_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 13 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day13 of Advent Of Code 2022
}
