% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{parse_day15_input}
\alias{parse_day15_input}
\title{Parsed Input to Day15 of Advent Of Code 2022}
\usage{
parse_day15_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 15 of AdventOfCode}
}
\value{
A list with a test row, a coordinate limit and a list of sensors, each with its position and the one of the closest beacon
}
\description{
Parsed Input to Day15 of Advent Of Code 2022
}
