% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day16.R
\name{parse_day16_input}
\alias{parse_day16_input}
\title{Parsed Input to Day16 of Advent Of Code 2022}
\usage{
parse_day16_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 16 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day16 of Advent Of Code 2022
}
