% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{parse_day17_input}
\alias{parse_day17_input}
\title{Parsed Input to Day17 of Advent Of Code 2022}
\usage{
parse_day17_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 17 of AdventOfCode}
}
\value{
A logical vector representing a jet pattern; TRUE means to the right and FALSE to the left
}
\description{
Parsed Input to Day17 of Advent Of Code 2022
}
