% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{parse_day18_input}
\alias{parse_day18_input}
\title{Parsed Input to Day18 of Advent Of Code 2022}
\usage{
parse_day18_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 18 of AdventOfCode}
}
\value{
A list of points in three dimensions representing occupied pixels
}
\description{
Parsed Input to Day18 of Advent Of Code 2022
}
