% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day1.R
\name{parse_day1_input}
\alias{parse_day1_input}
\title{Parsed Input to Day1 of Advent Of Code 2022}
\usage{
parse_day1_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 1 of AdventOfCode}
}
\value{
A list of one integer vector per backpack containing the individual calories per item
}
\description{
Parsed Input to Day1 of Advent Of Code 2022
}
