% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{parse_day20_input}
\alias{parse_day20_input}
\title{Parsed Input to Day20 of Advent Of Code 2022}
\usage{
parse_day20_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 20 of AdventOfCode}
}
\value{
An integer vector
}
\description{
Parsed Input to Day20 of Advent Of Code 2022
}
