% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day21.R
\name{parse_day21_input}
\alias{parse_day21_input}
\title{Parsed Input to Day21 of Advent Of Code 2022}
\usage{
parse_day21_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 21 of AdventOfCode}
}
\value{
Two lists, one for monkeys shouting numbers and one for operation monkeys
}
\description{
Parsed Input to Day21 of Advent Of Code 2022
}
