% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{parse_day24_input}
\alias{parse_day24_input}
\title{Parsed Input to Day24 of Advent Of Code 2022}
\usage{
parse_day24_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 24 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day24 of Advent Of Code 2022
}
