% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day25.R
\name{parse_day25_input}
\alias{parse_day25_input}
\title{Parsed Input to Day25 of Advent Of Code 2022}
\usage{
parse_day25_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 25 of AdventOfCode}
}
\value{
A list of fuel requirements encoded in SNAFU
}
\description{
Parsed Input to Day25 of Advent Of Code 2022
}
