% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{parse_day2_input}
\alias{parse_day2_input}
\title{Parsed Input to Day2 of Advent Of Code 2022}
\usage{
parse_day2_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 2 of AdventOfCode}
}
\value{
A tibble with columns \code{OpponentPlay} and \code{ResponseCode}
}
\description{
Parsed Input to Day2 of Advent Of Code 2022
}
