% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day4.R
\name{parse_day4_input}
\alias{parse_day4_input}
\title{Parsed Input to Day4 of Advent Of Code 2022}
\usage{
parse_day4_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 4 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day4 of Advent Of Code 2022
}
