% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day5.R
\name{parse_day5_input}
\alias{parse_day5_input}
\title{Parsed Input to Day5 of Advent Of Code 2022}
\usage{
parse_day5_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 5 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day5 of Advent Of Code 2022
}
