% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day6.R
\name{parse_day6_input}
\alias{parse_day6_input}
\title{Parsed Input to Day6 of Advent Of Code 2022}
\usage{
parse_day6_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 6 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day6 of Advent Of Code 2022
}
