% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day8.R
\name{parse_day8_input}
\alias{parse_day8_input}
\title{Parsed Input to Day8 of Advent Of Code 2022}
\usage{
parse_day8_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 8 of AdventOfCode}
}
\value{
A tibble representing a map of tree heights, each row representing a tree
}
\description{
Parsed Input to Day8 of Advent Of Code 2022
}
