% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{solve_day}
\alias{solve_day}
\title{Solve a Challenge from Advent of Code 2022}
\usage{
solve_day(input_string, day, part)
}
\arguments{
\item{input_string}{A character vector of length 1 containing the input for the specified problem in AdventOfCode}

\item{day}{An integer between 1 and 25 specifying the day of AdventOfCode}

\item{part}{The integer 1 or 2 specifying the part of the \code{day} of AdventOfCode}
}
\value{
A list with an elements 'solution', a string containing the solution, 'time_parsing', the time it took to parse the input, and 'time_execution', the time it took to compute the solution
}
\description{
Solve a Challenge from Advent of Code 2022
}
\examples{
solve_day("1000\n 2000\n 3000\n\n 4000", day = 1, part = 1)
}
