% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{solve_day10_part1}
\alias{solve_day10_part1}
\title{Solution to Day10 Part1 of Advent of Code 2022}
\usage{
solve_day10_part1(code)
}
\arguments{
\item{code}{A list of commands, each a list with arguments \code{type}, the command type, \code{argument}, the argument to the command, and \code{cyles}, the time it takes to execute the command}
}
\value{
The sum of the signal strength at each 20th cycle
}
\description{
Solution to Day10 Part1 of Advent of Code 2022
}
