% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{solve_day10_part2}
\alias{solve_day10_part2}
\title{Solution to Day10 Part2 of Advent of Code 2022}
\usage{
solve_day10_part2(code)
}
\arguments{
\item{code}{A list of commands, each a list with arguments \code{type}, the command type, \code{argument}, the argument to the command, and \code{cyles}, the time it takes to execute the command}
}
\value{
A string representing the display on the CRT monitor
}
\description{
Solution to Day10 Part2 of Advent of Code 2022
}
