% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{solve_day11_part1}
\alias{solve_day11_part1}
\title{Solution to Day11 Part1 of Advent of Code 2022}
\usage{
solve_day11_part1(monkeys)
}
\arguments{
\item{monkeys}{A list of monkeys}
}
\value{
The product of the number of items thrown by the two monkeys throwing the most in the first 20 rounds
}
\description{
Solution to Day11 Part1 of Advent of Code 2022
}
