% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{solve_day12_part2}
\alias{solve_day12_part2}
\title{Solution to Day12 Part2 of Advent of Code 2022}
\usage{
solve_day12_part2(height_map)
}
\arguments{
\item{height_map}{A list containing the start point, the end point and a height map}
}
\value{
The minimal distance between the end point and a loaction with the lowest height level
}
\description{
Solution to Day12 Part2 of Advent of Code 2022
}
