% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{solve_day13_part1}
\alias{solve_day13_part1}
\title{Solution to Day13 Part1 of Advent of Code 2022}
\usage{
solve_day13_part1(pairs)
}
\arguments{
\item{pairs}{A list of pairs of packets}
}
\value{
The sum of the indices of the pairs in the correct order
}
\description{
Solution to Day13 Part1 of Advent of Code 2022
}
