% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{solve_day14_part1}
\alias{solve_day14_part1}
\title{Solution to Day14 Part1 of Advent of Code 2022}
\usage{
solve_day14_part1(sand_cave)
}
\arguments{
\item{sand_cave}{A cave data structure containing the rock formations}
}
\value{
The number of sand packets coming to rest before one falls to the ground
}
\description{
Solution to Day14 Part1 of Advent of Code 2022
}
