% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{solve_day15_part1}
\alias{solve_day15_part1}
\title{Solution to Day15 Part1 of Advent of Code 2022}
\usage{
solve_day15_part1(input)
}
\arguments{
\item{input}{A list with a test row, a coordinate limit and a list of sensors, each with its position and the one of the closest beacon}
}
\value{
Points on the test row that cannot contain a beacon
}
\description{
Solution to Day15 Part1 of Advent of Code 2022
}
