% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day16.R
\name{solve_day16_part1}
\alias{solve_day16_part1}
\title{Solution to Day16 Part1 of Advent of Code 2022}
\usage{
solve_day16_part1(valves)
}
\arguments{
\item{valves}{A list of valves, each with its name, rate and connections to other valves}
}
\value{
The must pressure you can release in 30 minutes
}
\description{
Solution to Day16 Part1 of Advent of Code 2022
}
