% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{solve_day17_part2}
\alias{solve_day17_part2}
\title{Solution to Day17 Part2 of Advent of Code 2022}
\usage{
solve_day17_part2(jet_pattern)
}
\arguments{
\item{jet_pattern}{A logical vector representing a jet pattern; TRUE means to the right and FALSE to the left}
}
\value{
The height of the tower of the first 1,000,000,000,000 fallen rocks
}
\description{
Solution to Day17 Part2 of Advent of Code 2022
}
