% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{solve_day18_part1}
\alias{solve_day18_part1}
\title{Solution to Day18 Part1 of Advent of Code 2022}
\usage{
solve_day18_part1(pixels)
}
\arguments{
\item{pixels}{A list of points in three dimensions}
}
\value{
The number of the pixels' faces that do not immediately touch another pixel
}
\description{
Solution to Day18 Part1 of Advent of Code 2022
}
