% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{solve_day18_part2}
\alias{solve_day18_part2}
\title{Solution to Day18 Part2 of Advent of Code 2022}
\usage{
solve_day18_part2(pixels)
}
\arguments{
\item{pixels}{A list of points in three dimensions}
}
\value{
The number of the pixels' faces accessible from the outside
}
\description{
Solution to Day18 Part2 of Advent of Code 2022
}
