% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day1.R
\name{solve_day1_part1}
\alias{solve_day1_part1}
\title{Solution to Day1 Part1 of Advent of Code 2022}
\usage{
solve_day1_part1(backpacks)
}
\arguments{
\item{backpacks}{A list of one integer vector per backpack containing the individual calories per item}
}
\value{
The largest total calories in a backpack
}
\description{
Solution to Day1 Part1 of Advent of Code 2022
}
