% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day1.R
\name{solve_day1_part2}
\alias{solve_day1_part2}
\title{Solution to Day1 Part2 of Advent of Code 2022}
\usage{
solve_day1_part2(backpacks)
}
\arguments{
\item{backpacks}{A list of one integer vector per backpack containing the individual calories per item}
}
\value{
The sum of the three largest totals of calories in a backpacks
}
\description{
Solution to Day1 Part2 of Advent of Code 2022
}
