% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{solve_day20_part1}
\alias{solve_day20_part1}
\title{Solution to Day20 Part1 of Advent of Code 2022}
\usage{
solve_day20_part1(sequence)
}
\arguments{
\item{sequence}{An integer vector}
}
\value{
The sum of the 1000th, 2000th and 3000th element after 0 after shifting one round
}
\description{
Solution to Day20 Part1 of Advent of Code 2022
}
