% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day21.R
\name{solve_day21_part1}
\alias{solve_day21_part1}
\title{Solution to Day21 Part1 of Advent of Code 2022}
\usage{
solve_day21_part1(input)
}
\arguments{
\item{input}{TO DO}
}
\value{
TO DO
}
\description{
Solution to Day21 Part1 of Advent of Code 2022
}
