% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day22.R
\name{solve_day22_part1}
\alias{solve_day22_part1}
\title{Solution to Day22 Part1 of Advent of Code 2022}
\usage{
solve_day22_part1(input)
}
\arguments{
\item{input}{A list containing the borders of the map, the wall positions, by row and by columns and move instructions}
}
\value{
The final position and orientation after following the moves from the top-left encoded as an integer
}
\description{
Solution to Day22 Part1 of Advent of Code 2022
}
