% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day25.R
\name{solve_day25_part1}
\alias{solve_day25_part1}
\title{Solution to Day25 Part1 of Advent of Code 2022}
\usage{
solve_day25_part1(fuel_requirements)
}
\arguments{
\item{fuel_requirements}{A list of fuel fuel_requirements encoded in SNAFU}
}
\value{
The total fuel required encoded in SNAFU
}
\description{
Solution to Day25 Part1 of Advent of Code 2022
}
