% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2.R
\name{solve_day2_part1}
\alias{solve_day2_part1}
\title{Solution to Day2 Part1 of Advent of Code 2022}
\usage{
solve_day2_part1(strategy_guide)
}
\arguments{
\item{strategy_guide}{A tibble with columns \code{OpponentPlay} and \code{ResponseCode}}
}
\value{
The final player score if the guide is followed based on the part 1 assumptions
}
\description{
Solution to Day2 Part1 of Advent of Code 2022
}
