% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day3.R
\name{solve_day3_part1}
\alias{solve_day3_part1}
\title{Solution to Day3 Part1 of Advent of Code 2022}
\usage{
solve_day3_part1(rucksacks)
}
\arguments{
\item{rucksacks}{A list with one item per rucksack. Each rucksack is represented by a list of two character vectors containing the stored items, one per compartment.}
}
\value{
The total priority of the unique item types in each rucksack appearing in both compartments
}
\description{
Solution to Day3 Part1 of Advent of Code 2022
}
