% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day5.R
\name{solve_day5_part2}
\alias{solve_day5_part2}
\title{Solution to Day5 Part2 of Advent of Code 2022}
\usage{
solve_day5_part2(input)
}
\arguments{
\item{input}{A list containing an initial stack configuration and a list of moves.}
}
\value{
Top items after all moves are performed by the CrateMover 9001.
}
\description{
Solution to Day5 Part2 of Advent of Code 2022
}
