% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day7.R
\name{solve_day7_part1}
\alias{solve_day7_part1}
\title{Solution to Day7 Part1 of Advent of Code 2022}
\usage{
solve_day7_part1(commands)
}
\arguments{
\item{commands}{List of command line file system commands with output}
}
\value{
Sum of sizes of all directories with size <= 100000
}
\description{
Solution to Day7 Part1 of Advent of Code 2022
}
