% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day7.R
\name{solve_day7_part2}
\alias{solve_day7_part2}
\title{Solution to Day7 Part2 of Advent of Code 2022}
\usage{
solve_day7_part2(commands)
}
\arguments{
\item{commands}{List of command line file system commands with output}
}
\value{
Smallest directory size larger than the missing space to upgrade
}
\description{
Solution to Day7 Part2 of Advent of Code 2022
}
