% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day8.R
\name{solve_day8_part2}
\alias{solve_day8_part2}
\title{Solution to Day8 Part2 of Advent of Code 2022}
\usage{
solve_day8_part2(tree_map)
}
\arguments{
\item{tree_map}{A tibble representing a map of tree heights, each row representing a tree}
}
\value{
The number of visible trees on the map
}
\description{
Solution to Day8 Part2 of Advent of Code 2022
}
