% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day9.R
\name{solve_day9_part2}
\alias{solve_day9_part2}
\title{Solution to Day9 Part2 of Advent of Code 2022}
\usage{
solve_day9_part2(moves)
}
\arguments{
\item{moves}{A list of moves, each being a list with elements 'direction' out of ('U', 'D', 'L', 'R') and 'distance', an integer}
}
\value{
The number of points the tail of an 11 link rope touches if the head moves according to the input
}
\description{
Solution to Day9 Part2 of Advent of Code 2022
}
