% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{solve_day_function_name}
\alias{solve_day_function_name}
\title{Name of the function solving the specified problem in AdventOfCode}
\usage{
solve_day_function_name(day, part)
}
\arguments{
\item{day}{An integer between 1 and 25 specifying the day of AdventOfCode}

\item{part}{The integer 1 or 2 specifying the part of the \code{day} of AdventOfCode}
}
\value{
Name of the function solving the specified problem in AdventOfCode
}
\description{
The named function is the function solving based on parsed input.
}
