% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{solve_day_wo_validation}
\alias{solve_day_wo_validation}
\title{Solve a Challenge from Advent of Code 2022}
\usage{
solve_day_wo_validation(input_string, day, part)
}
\arguments{
\item{input_string}{A character vector of length 1 containing the input for the specified problem in AdventOfCode}

\item{day}{An integer between 1 and 25 specifying the day of AdventOfCode}

\item{part}{The integer 1 or 2 specifying the part of the \code{day} of AdventOfCode}
}
\value{
A string containing the solution
}
\description{
Solve a Challenge from Advent of Code 2022
}
\examples{
solve_day("1000\n 2000\n 3000\n\n 4000", day = 1, part = 1)
}
