% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_characters}
\alias{split_characters}
\title{Split Characters}
\usage{
split_characters(text)
}
\arguments{
\item{text}{A character vector of length 1}
}
\value{
A character vector containing each individual character
}
\description{
Split Characters
}
