% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_line_separated_blocks}
\alias{split_line_separated_blocks}
\title{Split Blocks of Lines}
\usage{
split_line_separated_blocks(text)
}
\arguments{
\item{text}{A character vector of length 1}
}
\value{
A character vector containing the blocks in \code{text} separated by an empty line
}
\description{
Split Blocks of Lines
}
