% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_lines}
\alias{split_lines}
\title{Split Lines}
\usage{
split_lines(text, remove_empty = TRUE)
}
\arguments{
\item{text}{A character vector of length 1}

\item{remove_empty}{A logical vector of length 1 specifying whether empty lines should be removed}
}
\value{
A character vector containing the line contents of each line of \code{text}.
}
\description{
Split Lines
}
