% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_character_grid}
\alias{to_character_grid}
\title{Split Lines}
\usage{
to_character_grid(text)
}
\arguments{
\item{text}{A character vector of length 1 whose lines have the same length}
}
\value{
A tibble with one row per character, excluding newlines, with an \code{Value} column containing the character, an \code{X} column containing the character position in the row and a \code{Y} column containing the line number of the character
}
\description{
Split Lines
}
