% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{validate_input_string}
\alias{validate_input_string}
\title{Validate AdventOfCode Input String}
\usage{
validate_input_string(input_string)
}
\arguments{
\item{input_string}{A string containing the input to an AdventOfCode problem.}
}
\value{
None
}
\description{
Basic validation for the input string to AdventOfCode problems.
Raises an error if the input is ot a character vector of length 1.
}
