% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{validate_problem_selection}
\alias{validate_problem_selection}
\title{Validate AdventOfCode Problem Specification}
\usage{
validate_problem_selection(day, part)
}
\arguments{
\item{day}{An integer specifying the day of AdventOfCode}

\item{part}{An integer specifying the day of AdventOfCode}
}
\value{
None
}
\description{
Raises an error if the provided values do not form a valid selection of an AdventOfCode problem.
}
