% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{parse_day10_input}
\alias{parse_day10_input}
\title{Parsed Input to Day10 of Advent Of Code 2022}
\usage{
parse_day10_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 10 of AdventOfCode}
}
\value{
A list of commands, each a list with arguments \code{type}, the command type, \code{argument}, the argument to the command, and \code{cyles}, the time it takes to execute the command
}
\description{
Parsed Input to Day10 of Advent Of Code 2022
}
