% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day22.R
\name{parse_day22_input}
\alias{parse_day22_input}
\title{Parsed Input to Day22 of Advent Of Code 2022}
\usage{
parse_day22_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 22 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day22 of Advent Of Code 2022
}
