% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day3.R
\name{parse_day3_input}
\alias{parse_day3_input}
\title{Parsed Input to Day3 of Advent Of Code 2022}
\usage{
parse_day3_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 3 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day3 of Advent Of Code 2022
}
