% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day9.R
\name{parse_day9_input}
\alias{parse_day9_input}
\title{Parsed Input to Day9 of Advent Of Code 2022}
\usage{
parse_day9_input(input_string)
}
\arguments{
\item{input_string}{Input string to day 9 of AdventOfCode}
}
\value{
TO DO
}
\description{
Parsed Input to Day9 of Advent Of Code 2022
}
