% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{solve_day14_part2}
\alias{solve_day14_part2}
\title{Solution to Day14 Part2 of Advent of Code 2022}
\usage{
solve_day14_part2(sand_cave)
}
\arguments{
\item{sand_cave}{A cave data structure containing the rock formations}
}
\value{
The number of sand packets coming to rest before one seals the drop spot
}
\description{
Solution to Day14 Part2 of Advent of Code 2022
}
