% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day21.R
\name{solve_day21_part1}
\alias{solve_day21_part1}
\title{Solution to Day21 Part1 of Advent of Code 2022}
\usage{
solve_day21_part1(monkeys)
}
\arguments{
\item{monkeys}{Two lists, one for monkeys shouting numbers and one for operation monkeys}
}
\value{
The number the root monkey will shout
}
\description{
Solution to Day21 Part1 of Advent of Code 2022
}
