% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day6.R
\name{solve_day6_part1}
\alias{solve_day6_part1}
\title{Solution to Day6 Part1 of Advent of Code 2022}
\usage{
solve_day6_part1(signal)
}
\arguments{
\item{signal}{A vector of single characters representing the signal}
}
\value{
Position of the last input in the first run of 4 distinct items
}
\description{
Solution to Day6 Part1 of Advent of Code 2022
}
